requirements.txt
fastapi
uvicorn
pydantic
google-genai
python-dotenv
requests